Use Option in metadata converter
Using an Option is a more natural way of expressing the intention. If I
give a ino, use it, if I don't then do something to generate one. At the
moment that means to take the ino of the existing file in SplitFS or use
the current timestamp in CatFS.

The Option object is preferred over conventions like using a special
value such as 0.