Fix misleading part in the README
The misleading part in the README said, that most cloud storage
providers do not support the upload of a single file. This is of course
rubbish. What I meant to say was, that they do not support the
concurrent upload of a single file, as in chunked upload.

This part is fixed now.