Notify main loop when filesystem is dropped
The filesystem implements the Drop trait now, which makes it possible to
run a function when the filesystem is unmouted in a way other than by
terminating the main loop (most prominently by using `umount` directly).

The previous situation was, when the filesystem was unmounted via
`umount`, then the main loop would hang infinitely, because there was no
way to notify the main loop. Now we send a quit signal when the
filesystem is dropped, so the main loop can exit normally.